requirements.txt
streamlit
moviepy
pillow
requests
edge-tts
openai
imageio
imageio-ffmpeg